@@ -5,6 +5,7 @@
   "dependencies": {
     "autoprefixer": "^6.4.1",
     "axios": "^0.16.2",
+    "babel-polyfill": "^6.26.0",
     "body-parser": "^1.16.0",
     "css-loader": "^0.26.1",
     "express": "^4.15.4",
@@ -28,14 +29,15 @@
   "devDependencies": {
     "babel-core": "^6.23.1",
     "babel-loader": "^6.3.2",
-    "babel-polyfill": "^6.23.0",
-    "babel-preset-env": "^1.1.8",
-    "babel-preset-latest": "^6.22.0",
-    "babel-preset-react": "^6.23.0",
+    "babel-preset-env": "^1.6.0",
+    "babel-preset-latest": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-2": "^6.24.1",
+    "babel-preset-stage-3": "^6.24.1",
     "json-loader": "^0.5.7",
     "node-sass": "^4.5.3",
     "sass-loader": "^6.0.6",
-    "webpack": "^2.2.1",
+    "webpack": "^2.7.0",
     "webpack-dev-middleware": "^1.12.0",
     "webpack-dev-server": "^2.4.1"
   }
