@@ -1,50 +1,9 @@
 import React, { Component } from 'react';
 
-var SpotifyWebApi = require('spotify-web-api-node');
-var spotifyApi = new SpotifyWebApi();
-
+import Tracklist from './Tracklist';
 import Playlists from "./Playlists";
 
 class Home extends Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            token: "",
-            username: "",
-            playlists: ""
-        }
-    }
-    componentDidMount() {
-        if (this.props.location.hash) {
-            let newHash = this.props.location.hash.slice(14, -44);
-            this.setState({ token: newHash });
-        }
-        setTimeout(() => {
-            this.getMe();
-        }, 100)
-        
-    }
-    getPlaylists() {
-        spotifyApi.getUserPlaylists(this.state.username)
-            .then((data) => {
-                this.setState({playlists: data.body.items});
-            }, function (err) {
-                console.log('Something went wrong!', err);
-            });
-    }
-    getMe() {
-        spotifyApi.setAccessToken(this.state.token);
-
-        spotifyApi.getMe()
-            .then((data) =>  {
-                this.setState({username: data.body.id});
-            }, function (err) {
-                console.log('Something went wrong!', err);
-            });
-        setTimeout(() => {
-            this.getPlaylists();
-        }, 300)
-    }
     query() {
         spotifyApi.setAccessToken(this.state.token);
 
@@ -81,8 +40,11 @@ class Home extends Component {
         return (
             <div>
                 <div className="columns">
-                    <div className="column is-2 playlists-menu">
-                        <Playlists playlists={this.state.playlists}/>
+                    <div className="column playlists-menu">
+                        <Playlists onClickPlaylist={this.props.onClickPlaylist} playlists={this.props.playlists} />
+                    </div>
+                    <div className="column is-8 tracklist">
+                        <Tracklist tracklist={this.props.tracklist} />
                     </div>
                 </div>
 
