@@ -1,13 +1,18 @@
 import React, { Component } from 'react';
 import { Link } from "react-router";
+import { browserHistory } from 'react-router';
 
 class Nav extends Component {
+    onLogout = () => {
+        localStorage.removeItem('token');
+        browserHistory.replace('/');
+    }
 	render() {
 		return (
 			<nav className="navbar ">
 				<div className="navbar-brand">
 					<a className="navbar-item" href="/">
-						<img src="http://bulma.io/images/bulma-logo.png" alt="Bulma: a modern CSS framework based on Flexbox"  />
+						<p>SoundIze</p>
 					</a>
 
 					<a className="navbar-item is-hidden-desktop" href="https://github.com/jgthms/bulma" target="_blank">
@@ -31,10 +36,10 @@ class Nav extends Component {
 
 				<div id="navMenubd-example" className="navbar-menu">
 					<div className="navbar-start">
-						<Link className="navbar-item " to="/hem">
+						<Link className="navbar-item " to="/app/stream">
 							Hem
       					</Link>
-						<Link className="navbar-item " to="/test">
+						<Link className="navbar-item " to="/">
 							Test
      					</Link>
 					</div>
@@ -43,29 +48,12 @@ class Nav extends Component {
 						<div className="navbar-item">
 							<div className="field is-grouped">
 								<p className="control">
-									<a className="bd-tw-button button"
-										data-social-network="Twitter"
-										data-social-action="tweet"
-										data-social-target="http://bulma.io"
-										target="_blank"
-										href="https://twitter.com/intent/tweet?text=Bulma: a modern CSS framework based on Flexbox&hashtags=bulmaio&url=http://bulma.io&via=jgthms">
-										<span className="icon">
-											<i className="fa fa-twitter"></i>
-										</span>
-										<span>
-											Tweet
+									<a className="bd-tw-button button">
+										<span onClick={() => this.onLogout()}>
+											Logga ut
   										</span>
 									</a>
-
-								</p>
-								<p className="control">
-									<a className="button is-primary" href="https://github.com/jgthms/bulma/archive/0.5.1.zip">
-										<span className="icon">
-											<i className="fa fa-download"></i>
-										</span>
-										<span>Download</span>
-									</a>
-								</p>
+								</p>						
 							</div>
 						</div>
 					</div>
